Add a license notice on top of each file

+++ src/Settings.cpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 SemkiShow
+//
+// SPDX-License-Identifier: GPL-3.0-only
+
 #include "Settings.hpp"
 #include <fstream>
 #include <vector>
@@ -10,7 +14,7 @@ std::vector<std::string> Split(std::string input, char delimiter = ' ')
     std::vector<std::string> output;
     output.push_back("");
     int index = 0;
-    for (int i = 0; i < input.size(); i++)
+    for (size_t i = 0; i < input.size(); i++)
     {
         if (input[i] == delimiter)
         {
